Extract workspace worktree APIs
