+++ Tests/QuillCodeParityTests/ParityWorkspaceProjectGateTests.swift
@@ -141,6 +141,7 @@ final class ParityWorkspaceProjectGateTests: QuillCodeParityTestCase {
 
     func testWorkspaceModelDelegatesWorktreeOpenRecords() throws {
         let modelText = try Self.appSourceText(named: "WorkspaceModel.swift")
+        let worktreeExtensionText = try Self.appSourceText(named: "WorkspaceModelWorktrees.swift")
         let requestsText = try Self.appSourceText(named: "WorkspaceWorktreeRequests.swift")
         let engineText = try Self.appSourceText(named: "WorkspaceWorktreeOpenEngine.swift")
         let plannerText = try Self.appSourceText(named: "WorkspaceWorktreeToolCallPlanner.swift")
@@ -155,17 +156,22 @@ final class ParityWorkspaceProjectGateTests: QuillCodeParityTestCase {
         XCTAssertTrue(plannerText.contains("static func create"), "Worktree create tool calls should be directly testable.")
         XCTAssertTrue(plannerText.contains("static func remove"), "Worktree remove tool calls should be directly testable.")
         XCTAssertTrue(plannerText.contains("static func prune"), "Worktree prune tool calls should be directly testable.")
-        XCTAssertTrue(modelText.contains("WorkspaceWorktreeToolCallPlanner.create"), "WorkspaceModel should delegate worktree create tool-call construction.")
-        XCTAssertTrue(modelText.contains("WorkspaceWorktreeToolCallPlanner.remove"), "WorkspaceModel should delegate worktree remove tool-call construction.")
-        XCTAssertTrue(modelText.contains("WorkspaceWorktreeToolCallPlanner.prune"), "WorkspaceModel should delegate worktree prune tool-call construction.")
-        XCTAssertTrue(modelText.contains("WorkspaceWorktreeOpenEngine.localThread"), "WorkspaceModel should delegate local worktree handoff records.")
-        XCTAssertTrue(modelText.contains("WorkspaceWorktreeOpenEngine.remoteThread"), "WorkspaceModel should delegate SSH Remote worktree handoff records.")
-        XCTAssertTrue(modelText.contains("openCreatedWorktreeThread"), "WorkspaceModel should share selected-thread persistence for local and remote worktrees.")
-        XCTAssertFalse(modelText.contains("ToolDefinition.gitWorktreeCreate.name"), "WorkspaceModel should not own worktree create tool-call details.")
-        XCTAssertFalse(modelText.contains("ToolDefinition.gitWorktreeRemove.name"), "WorkspaceModel should not own worktree remove tool-call details.")
-        XCTAssertFalse(modelText.contains("ToolDefinition.gitWorktreePrune.name"), "WorkspaceModel should not own worktree prune tool-call details.")
-        XCTAssertFalse(modelText.contains("title: \"Worktree:"), "WorkspaceModel should not own worktree thread title copy.")
-        XCTAssertFalse(modelText.contains("Opened remote worktree `"), "WorkspaceModel should not own remote worktree transcript copy.")
-        XCTAssertFalse(modelText.contains("Opened worktree `"), "WorkspaceModel should not own local worktree transcript copy.")
+        XCTAssertTrue(worktreeExtensionText.contains("extension QuillCodeWorkspaceModel"), "Worktree APIs should live in a focused WorkspaceModel extension.")
+        XCTAssertTrue(worktreeExtensionText.contains("WorkspaceWorktreeToolCallPlanner.create"), "Worktree extension should delegate worktree create tool-call construction.")
+        XCTAssertTrue(worktreeExtensionText.contains("WorkspaceWorktreeToolCallPlanner.remove"), "Worktree extension should delegate worktree remove tool-call construction.")
+        XCTAssertTrue(worktreeExtensionText.contains("WorkspaceWorktreeToolCallPlanner.prune"), "Worktree extension should delegate worktree prune tool-call construction.")
+        XCTAssertTrue(worktreeExtensionText.contains("WorkspaceWorktreeOpenEngine.localThread"), "Worktree extension should delegate local worktree handoff records.")
+        XCTAssertTrue(worktreeExtensionText.contains("WorkspaceWorktreeOpenEngine.remoteThread"), "Worktree extension should delegate SSH Remote worktree handoff records.")
+        XCTAssertTrue(worktreeExtensionText.contains("openCreatedWorktreeThread"), "Worktree extension should share selected-thread persistence for local and remote worktrees.")
+        XCTAssertFalse(modelText.contains("public func createWorktree"), "WorkspaceModel.swift should not own worktree create APIs.")
+        XCTAssertFalse(modelText.contains("public func openWorktree"), "WorkspaceModel.swift should not own worktree open APIs.")
+        XCTAssertFalse(modelText.contains("public func removeWorktree"), "WorkspaceModel.swift should not own worktree remove APIs.")
+        XCTAssertFalse(modelText.contains("public func pruneWorktrees"), "WorkspaceModel.swift should not own worktree prune APIs.")
+        XCTAssertFalse(worktreeExtensionText.contains("ToolDefinition.gitWorktreeCreate.name"), "Worktree extension should not own worktree create tool-call details.")
+        XCTAssertFalse(worktreeExtensionText.contains("ToolDefinition.gitWorktreeRemove.name"), "Worktree extension should not own worktree remove tool-call details.")
+        XCTAssertFalse(worktreeExtensionText.contains("ToolDefinition.gitWorktreePrune.name"), "Worktree extension should not own worktree prune tool-call details.")
+        XCTAssertFalse(worktreeExtensionText.contains("title: \"Worktree:"), "Worktree extension should not own worktree thread title copy.")
+        XCTAssertFalse(worktreeExtensionText.contains("Opened remote worktree `"), "Worktree extension should not own remote worktree transcript copy.")
+        XCTAssertFalse(worktreeExtensionText.contains("Opened worktree `"), "Worktree extension should not own local worktree transcript copy.")
     }
 }
